test(Coordinate): Coordinate 테스트 추가
- 좌표를 180도 돌렸을 경우 올바른 좌표를 가져오는지에 대한 테스트
- 패키지 구조 변경
