+++ src/test/java/domain/position/PositionTest.java
@@ -1,8 +1,6 @@
-package position;
+package domain.position;
 
 import domain.board.Direction;
-import domain.position.Coordinate;
-import domain.position.Position;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
